feat/ terminando algumas adições
