+++ semana17/revisao/src/data/selectActorByName.ts
@@ -1,0 +1,9 @@
+import connection from '../connection';
+
+export default selectActorByName = async (name: string): Promise<any> => {
+    const result = await connection.raw(`
+      SELECT * FROM Actor WHERE name = '${name}'
+    `)
+  
+      return result[0][0]
+  }
