+++ semana17/revisao/src/endpoints/getActorByName.ts
@@ -1,0 +1,13 @@
+import { Response, Request } from "express";
+import sele
+
+export default async function getActorByName (
+    req: Request,
+    res: Response
+) {
+    try {
+        const actor = await selectActorByName 
+    } catch (error:any | undefined) {
+        
+    }
+}
