+++ semana17/revisao/src/index.ts
@@ -1,0 +1,21 @@
+import connection from "./connection";
+import express from "express";
+import cors from "cors";
+import { AddressInfo } from "net";
+
+const app = express();
+
+app.use(express.json());
+app.use(cors());
+
+const server = app.listen(process.env.PORT || 3003, () => {
+  if (server) {
+    const address = server.address() as AddressInfo;
+    console.log(`Server is running in http://localhost:${address.port}`);
+  } else {
+    console.error(`Failure upon starting server.`);
+  }
+});
+
+app.get ("/", getActorByName)
+
